Evitar consultar a la base de datos apenas cargue railway

+++ app/Providers/AuthServiceProvider.php
@@ -8,6 +8,7 @@ use Carbon\Carbon;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Pluralizer;
 
 class AuthServiceProvider extends ServiceProvider
@@ -34,6 +35,9 @@ class AuthServiceProvider extends ServiceProvider
             return 'App\\Policies\\Api\\v1\\'. Pluralizer::singular(class_basename($model)).'Policy';
         });
 
-        AuthService::RemoveExpiredTokensAction();
+        if (Schema::hasTable('personal_access_tokens')) {
+            AuthService::RemoveExpiredTokensAction();
+        }
+
     }
 }
